+++ pages/api/premintCheck/[address].ts
@@ -4,16 +4,24 @@ import { Signature } from 'ethers'
 
 import { METABOT_BASE_API_URL } from 'utils/constants'
 import logbookMongoose from 'utils/logbookMongoose'
+import { LogData, logError, logSuccess } from 'utils/logging'
 import { fetcher } from 'utils/requests'
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-    try {
-        const address = req.query.address as string
+    const address = req.query.address as string
 
+    const logData: LogData = {
+        function_name: 'premintCheck',
+        message: `begin`,
+        wallet_address: address,
+    }
+    try {
         type PremintCheckResponse = {
             allowlist: boolean
             signature: Signature | null
         }
+
+        logData.third_party_name = 'metabot'
 
         const { allowlist, signature } = (await fetcher(
             `${METABOT_BASE_API_URL}premintCheck/${address}`,
@@ -39,9 +47,10 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             response.errorCode = 2
         }
 
+        logSuccess(logData, response.message || 'success')
         return res.status(200).json(response)
     } catch (error) {
-        console.log(error)
+        logError(logData, error)
         res.status(500).json({ error: error.message })
     }
 }
